Part 11 - Java Part

+++ src/main/java/com/aivlev/vcp/service/impl/CategoryServiceImpl.java
@@ -1,5 +1,7 @@
 package com.aivlev.vcp.service.impl;
 
+import com.aivlev.vcp.exception.DuplicateEntityException;
+import com.aivlev.vcp.exception.ModelNotFoundException;
 import com.aivlev.vcp.model.Category;
 import com.aivlev.vcp.repository.storage.CategoryRepository;
 import com.aivlev.vcp.service.CategoryService;
@@ -24,22 +26,38 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public Category saveCategory(Category category) {
-        return categoryRepository.save(category);
-    }
-
-    @Override
     public Category findCategory(String id) {
-        return categoryRepository.findOne(id);
+        Category category = categoryRepository.findOne(id);
+        if(null != category){
+            return category;
+        } else {
+            throw new ModelNotFoundException("Category not found");
+        }
     }
 
     @Override
-    public Category updateCategory(Category category) {
-        return categoryRepository.save(category);
+    public void createOrUpdate(String id, Category category) {
+        if(null != id){
+            Category categoryFromDB = categoryRepository.findOne(id);
+            if(categoryFromDB == null){
+                throw new ModelNotFoundException("Category not found");
+            }
+        }
+        try {
+            categoryRepository.save(category);
+        } catch (Exception ex){
+            throw new DuplicateEntityException("Category is exists");
+        }
+
     }
 
     @Override
     public void deleteCategory(String id) {
-        categoryRepository.delete(id);
+        Category categoryFromDB = categoryRepository.findOne(id);
+        if(null != categoryFromDB){
+            categoryRepository.delete(id);
+        } else {
+            throw new ModelNotFoundException("Category not found");
+        }
     }
 }
